--- src/main/java/com/project/team11_tabling/booking/Service/BookingService.java
@@ -1,5 +1,0 @@
-package com.project.team11_tabling.booking.Service;
-
-public interface BookingService {
-
-}
