--- src/main/java/com/project/team11_tabling/booking/Service/BookingServiceImpl.java
@@ -1,8 +1,0 @@
-package com.project.team11_tabling.booking.Service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class BookingServiceImpl implements BookingService{
-
-}
