--- src/main/java/com/project/team11_tabling/booking/controller/BookingController.java
@@ -1,8 +1,0 @@
-package com.project.team11_tabling.booking.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class BookingController {
-
-}
