--- src/main/java/com/project/team11_tabling/booking/dto/BookingRequest.java
@@ -1,5 +1,0 @@
-package com.project.team11_tabling.booking.dto;
-
-public class BookingRequest {
-
-}
