--- src/main/java/com/project/team11_tabling/booking/repository/BookingRepository.java
@@ -1,8 +1,0 @@
-package com.project.team11_tabling.booking.repository;
-
-import com.project.team11_tabling.booking.Entity.Booking;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BookingRepository extends JpaRepository<Booking, Long> {
-
-}
