+++ src/main/java/com/project/team11_tabling/domain/booking/Service/BookingService.java
@@ -1,0 +1,5 @@
+package com.project.team11_tabling.domain.booking.Service;
+
+public interface BookingService {
+
+}
