+++ src/main/java/com/project/team11_tabling/domain/booking/Service/BookingServiceImpl.java
@@ -1,0 +1,8 @@
+package com.project.team11_tabling.domain.booking.Service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class BookingServiceImpl implements BookingService{
+
+}
