+++ src/main/java/com/project/team11_tabling/domain/booking/controller/BookingController.java
@@ -1,0 +1,8 @@
+package com.project.team11_tabling.domain.booking.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class BookingController {
+
+}
