+++ src/main/java/com/project/team11_tabling/domain/booking/dto/BookingRequest.java
@@ -1,0 +1,5 @@
+package com.project.team11_tabling.domain.booking.dto;
+
+public class BookingRequest {
+
+}
