+++ src/main/java/com/project/team11_tabling/domain/booking/repository/BookingRepository.java
@@ -1,0 +1,8 @@
+package com.project.team11_tabling.domain.booking.repository;
+
+import com.project.team11_tabling.domain.booking.Entity.Booking;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BookingRepository extends JpaRepository<Booking, Long> {
+
+}
