+++ src/main/java/com/project/team11_tabling/domain/user/controller/UserController.java
@@ -1,0 +1,14 @@
+package com.project.team11_tabling.domain.user.controller;
+
+import com.project.team11_tabling.domain.user.service.UserService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("")
+@RequiredArgsConstructor
+public class UserController {
+  private final UserService userService;
+
+}
