+++ src/main/java/com/project/team11_tabling/domain/user/dto/LoginRequestDto.java
@@ -1,0 +1,11 @@
+package com.project.team11_tabling.domain.user.dto;
+
+import lombok.Getter;
+
+@Getter
+public class LoginRequestDto {
+
+  private String email;
+  private String password;
+  private Long phoneNumber;
+}
