+++ src/main/java/com/project/team11_tabling/domain/user/dto/ModifyPasswordDto.java
@@ -1,0 +1,11 @@
+package com.project.team11_tabling.domain.user.dto;
+
+import lombok.Getter;
+
+@Getter
+public class ModifyPasswordDto {
+
+  private String password;
+  private String changePassword;
+  private String changePasswordConfirm;
+}
