+++ src/main/java/com/project/team11_tabling/domain/user/dto/UserRequestDto.java
@@ -1,0 +1,12 @@
+package com.project.team11_tabling.domain.user.dto;
+
+import lombok.Getter;
+
+@Getter
+public class UserRequestDto {
+
+  private String email;
+  private String username;
+  private String password;
+  private Long phoneNumber;
+}
