+++ src/main/java/com/project/team11_tabling/domain/user/service/UserService.java
@@ -1,0 +1,13 @@
+package com.project.team11_tabling.domain.user.service;
+
+import com.project.team11_tabling.domain.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserService {
+
+  private final UserRepository userRepository;
+
+}
