+++ src/main/java/com/project/team11_tabling/domain/wish/controller/WishController.java
@@ -1,0 +1,8 @@
+package com.project.team11_tabling.domain.wish.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class WishController {
+
+}
