+++ src/main/java/com/project/team11_tabling/domain/wish/dto/Response.java
@@ -1,0 +1,5 @@
+package com.project.team11_tabling.domain.wish.dto;
+
+public class Response {
+
+}
