+++ src/main/java/com/project/team11_tabling/domain/wish/repository/WishRepository.java
@@ -1,0 +1,8 @@
+package com.project.team11_tabling.domain.wish.repository;
+
+import com.project.team11_tabling.domain.wish.entity.Wish;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface WishRepository extends JpaRepository<Wish, Long> {
+
+}
