+++ src/main/java/com/project/team11_tabling/domain/wish/service/WishService.java
@@ -1,0 +1,8 @@
+package com.project.team11_tabling.domain.wish.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class WishService {
+
+}
