+++ src/main/java/com/project/team11_tabling/global/jwt/config/WebSecurityConfig.java
@@ -1,0 +1,5 @@
+package com.project.team11_tabling.global.jwt.config;
+
+public class WebSecurityConfig {
+
+}
