+++ src/main/java/com/project/team11_tabling/global/jwt/security/UserDetailsImpl.java
@@ -1,0 +1,5 @@
+package com.project.team11_tabling.global.jwt.security;
+
+public class UserDetailsImpl {
+
+}
