--- src/main/java/com/project/team11_tabling/jwt/JwtUtil.java
@@ -1,5 +1,0 @@
-package com.project.team11_tabling.jwt;
-
-public class JwtUtil {
-
-}
