--- src/main/java/com/project/team11_tabling/jwt/config/WebSecurityConfig.java
@@ -1,5 +1,0 @@
-package com.project.team11_tabling.jwt.config;
-
-public class WebSecurityConfig {
-
-}
