--- src/main/java/com/project/team11_tabling/jwt/security/JwtAuthorizationFilter.java
@@ -1,5 +1,0 @@
-package com.project.team11_tabling.jwt.security;
-
-public class JwtAuthorizationFilter {
-
-}
