--- src/main/java/com/project/team11_tabling/user/dto/LoginRequestDto.java
@@ -1,11 +1,0 @@
-package com.project.team11_tabling.user.dto;
-
-import lombok.Getter;
-
-@Getter
-public class LoginRequestDto {
-
-  private String email;
-  private String password;
-  private Long phoneNumber;
-}
