--- src/main/java/com/project/team11_tabling/user/dto/ModifyPasswordDto.java
@@ -1,11 +1,0 @@
-package com.project.team11_tabling.user.dto;
-
-import lombok.Getter;
-
-@Getter
-public class ModifyPasswordDto {
-
-  private String password;
-  private String changePassword;
-  private String changePasswordConfirm;
-}
