--- src/main/java/com/project/team11_tabling/user/dto/UserRequestDto.java
@@ -1,12 +1,0 @@
-package com.project.team11_tabling.user.dto;
-
-import lombok.Getter;
-
-@Getter
-public class UserRequestDto {
-
-  private String email;
-  private String username;
-  private String password;
-  private Long phoneNumber;
-}
