--- src/main/java/com/project/team11_tabling/user/dto/UserResponseDto.java
@@ -1,17 +1,0 @@
-package com.project.team11_tabling.user.dto;
-
-import lombok.Getter;
-
-@Getter
-public class UserResponseDto {
-
-  private String email;
-  private String username;
-
-  public UserResponseDto(String email, String username) {
-    this.email = email;
-    this.username = username;
-  }
-}
-
-
