--- src/main/java/com/project/team11_tabling/user/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.project.team11_tabling.user.repository;
-
-
-import com.project.team11_tabling.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-}
