--- src/main/java/com/project/team11_tabling/user/service/UserService.java
@@ -1,13 +1,0 @@
-package com.project.team11_tabling.user.service;
-
-import com.project.team11_tabling.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class UserService {
-
-  private final UserRepository userRepository;
-
-}
