--- src/main/java/com/project/team11_tabling/wish/controller/WishController.java
@@ -1,8 +1,0 @@
-package com.project.team11_tabling.wish.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class WishController {
-
-}
