--- src/main/java/com/project/team11_tabling/wish/dto/Response.java
@@ -1,5 +1,0 @@
-package com.project.team11_tabling.wish.dto;
-
-public class Response {
-
-}
