--- src/main/java/com/project/team11_tabling/wish/repository/WishRepository.java
@@ -1,8 +1,0 @@
-package com.project.team11_tabling.wish.repository;
-
-import com.project.team11_tabling.wish.entity.Wish;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface WishRepository extends JpaRepository<Wish, Long> {
-
-}
