--- src/main/java/com/project/team11_tabling/wish/service/WishService.java
@@ -1,8 +1,0 @@
-package com.project.team11_tabling.wish.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class WishService {
-
-}
